clean up Dockerfile for a working build
Both the `golang` and `busybox` containers use the `latest` tag.
This gets us an UTD golang version as well as a `uclibc` busybox container
which should result in a (slightly) smaller deployment container.

the `RUN` commands have been broken up to improve caching.
the short form arguments on `curl` calls have been converted into
long-form to improve readability.

some comments added to improve maintainability.

golint is now installed in the container. It was used in the `Makefile`
but not installed in the container... creating errors on build